requirements.txt
Flask==2.3.2
tensorflow==2.20.0
numpy
pandas
python-dotenv
beautifulsoup4
requests
